Extract head seo method in a separate function to be used independently
